# backend-dotnet/api/openapi.yaml
openapi: 3.1.0
info:
  title: Claims API (life claim slice)
  version: 0.2.0
  summary: Individual life, disability income and annuity claims. This version covers one life claim from intake to closed (decision, payment run, close).
  description: |
    Derived from the React mock's `web/src/api` modules (their doc comments carry the verbs). Where the mock's shape is
    unsuitable for a real API the difference is deliberate; the full list is in `backend/README.md` ("Deviations from the mock").

    **Conventions**

    * **Ids** are UUIDs. Claims also have a human `claimNumber` (`L-26-043310`), unique, filterable, never used in paths.
    * **Money** is `{ "amount": "200000.00", "currency": "USD" }`: a decimal string with at most 2 decimals, never a JSON number.
    * **Timestamps** are RFC 3339 in UTC (`2026-09-25T15:03:00Z`). Business dates (date of death, issue date) are plain dates.
    * **Enum values** are snake_case, the same strings the database stores (and the deadline kinds in the architecture doc).
    * **Idempotency**: POSTs that create something take an `Idempotency-Key` header (8 to 128 characters). A repeat with the
      same key and body returns the resource it created (200, `Idempotent-Replayed: true`); the same key with a different body is 409.
    * **Optimistic concurrency**: every mutable resource has a `version`, returned as its `ETag` (`"3"`). State-changing calls on
      an existing resource require `If-Match`: missing is 428, stale is 412.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /claims/{claimId}/requirements:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Requirements]
      operationId: listClaimRequirements
      summary: Requirements on a claim
      description: Mock `listRequirements(claimId)`.
      x-slice-status: implemented
      responses:
        '200':
          description: All requirements, oldest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementList' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requirements/{requirementId}:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    get:
      tags: [Requirements]
      operationId: getRequirement
      summary: Get a requirement
      description: One requirement, with its ETag for a later accept or waive.
      x-slice-status: implemented
      responses:
        '200':
          description: The requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requirements/{requirementId}:accept:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    post:
      tags: [Requirements]
      operationId: acceptRequirement
      summary: Accept a requirement
      description: |
        Postgres only, one transaction: the requirement becomes `accepted`, its waiting follow-up deadline rows close without
        firing, and if it was the last one, proof of loss is complete (status `in_review`, `decision_due` and `review_target`
        rows written, a work item for the owner). Mock: `markReceived` / `satisfyRequirement`. The event workflow that would call
        this when a document is classified is not built.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                satisfiedBy: { type: string, description: 'What satisfied it, shown on the requirement.' }
      responses:
        '200':
          description: The accepted requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /requirements/{requirementId}:waive:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    post:
      tags: [Requirements]
      operationId: waiveRequirement
      summary: Waive a requirement, with a reason
      description: Mock `waiveRequirement`. Same effects as accept. A reason is required (422 `reason_required`).
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1 }
      responses:
        '200':
          description: The waived requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /claims/{claimId}/deadlines:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Deadlines]
      operationId: listClaimDeadlines
      summary: Deadline rows for a claim
      description: Mock `LifeFlow.deadlines` in `GET /claims/{id}/workflow`, split into its own resource.
      x-slice-status: implemented
      responses:
        '200':
          description: All rows, oldest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeadlineList' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deadlines:
    get:
      tags: [Deadlines]
      operationId: listDeadlines
      summary: Deadlines across claims (operations view)
      description: What is open, overdue, or dispatched and stuck. Ordered by `dueAt`.
      x-slice-status: implemented
      parameters:
        - { name: state, in: query, schema: { $ref: '#/components/schemas/DeadlineState' } }
        - { name: kind, in: query, schema: { $ref: '#/components/schemas/DeadlineKind' } }
        - { name: dueBefore, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: Deadline rows.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeadlineList' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /deadlines/{deadlineId}:
    parameters: [{ $ref: '#/components/parameters/DeadlineId' }]
    get:
      tags: [Deadlines]
      operationId: getDeadline
      summary: Get a deadline row
      description: One deadline row, with its ETag for a later extension.
      x-slice-status: implemented
      responses:
        '200':
          description: The row.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deadline' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deadlines/{deadlineId}:extend:
    parameters: [{ $ref: '#/components/parameters/DeadlineId' }]
    post:
      tags: [Deadlines]
      operationId: extendDeadline
      summary: Extend an open deadline, with a reason
      description: A row update, not a new timer. `dueAt` moves, `originalDueAt` stays, the history records why.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newDueAt, reason]
              properties:
                newDueAt: { type: string, format: date-time }
                reason: { type: string, minLength: 1 }
      responses:
        '200':
          description: The extended row.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deadline' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /claims/{claimId}/letters:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Letters]
      operationId: listClaimLetters
      summary: Outbound letters and messages
      description: Mock `GET /claims/{id}/communications`, outbound only. Calls and inbound items are history events and documents.
      x-slice-status: implemented
      responses:
        '200':
          description: Newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Letter' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
  /claims/{claimId}/history:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [History]
      operationId: listClaimHistory
      summary: Append-only history, newest first
      description: Mock `listHistory`. Events are never edited or deleted.
      x-slice-status: implemented
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of events.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/HistoryEvent' } }
                  nextCursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /claims/{claimId}/workflow-runs:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Workflow runs]
      operationId: listClaimWorkflowRuns
      summary: What each workflow run did
      description: |
        Mock `LifeFlow.runs`. Written by the workflows' own activities, so the record and the state change it describes
        commit together. `swimlane` rows and `serviceLevels` are projections and are not part of this resource.
      x-slice-status: implemented
      responses:
        '200':
          description: Runs, oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/WorkflowRun' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'The intended scheme (org identity provider). NOT enforced yet: the service is unauthenticated and trusts X-Actor. See the README, not built yet.'
  headers:
    ETag:
      description: The resource's version, quoted. Send it back as `If-Match`.
      schema: { type: string, examples: ['"3"'] }
  parameters:
    ClaimId: { name: claimId, in: path, required: true, schema: { type: string, format: uuid } }
    RequirementId: { name: requirementId, in: path, required: true, schema: { type: string, format: uuid } }
    DeadlineId: { name: deadlineId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The ETag from the last GET of this resource.
      schema: { type: string, examples: ['"3"'] }
...
    Actor:
      name: X-Actor
      in: header
      required: false
      description: Development placeholder for the acting user until authentication is built.
      schema: { type: string }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque; the `nextCursor` of the previous page.
      schema: { type: string }
  responses:
    BadRequest:
      description: The request could not be read (malformed JSON, bad header or cursor).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    NotFound:
      description: No such resource.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: The request conflicts with the resource's state (`invalid_state`, `idempotency_key_reuse`, `deadline_not_open`, `claim_not_ready_to_decide`, `not_awaiting_approval`, `item_in_run`, ...).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionFailed:
      description: '`If-Match` is stale (`version_conflict`). Fetch the resource again.'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionRequired:
      description: '`If-Match` is required (`precondition_required`).'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unprocessable:
      description: Well formed but not valid (`validation_failed` with field errors, or a business-rule code such as `payee_shares_invalid`).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Stable, machine-readable. Clients branch on this, not on `detail`.
          examples: [validation_failed, version_conflict, idempotency_key_reuse]
        errors:
          type: array
          description: Field-level details for `validation_failed`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    DeadlineKind:
      type: string
      description: |
        Only `requirement_follow_up` has a workflow (the dispatcher fires it). Every other kind is a row that something else closes, or a target the
        nightly overdue check raises to a person when it is late: `document_review_by` (1 business day for the examiner to review a document the
        rules could not accept; closed by the review) and `bank_details_by` (a payee's new bank details after a returned payment, 10 days; closed by
        `update-payment-method`) are display/target rows with no automation of their own. `payment_due` also carries the "reissue a returned
        payment" target (`sla: reissue`), closed by the payment run.
      enum: [acknowledge_by, forms_by, first_contact_by, status_letter, requirement_follow_up, decision_due, review_target, payment_due, document_review_by, bank_details_by]
    DeadlineState:
      type: string
      description: open, then dispatched when the dispatcher starts its workflow, then done or skipped.
      enum: [open, dispatched, done, skipped]
    RequirementState:
      type: string
      enum: [requested, received, accepted, not_enough, waived, expired]
...
    Requirement:
      type: object
      required: [id, claimId, key, name, purpose, from, state, requestedAt, followUpDays, followUpCount, version]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        benefitLineId: { type: [string, 'null'], format: uuid }
        key: { type: string, enum: [certificate, statement, primary_died_first, report, amended_certificate] }
        name: { type: string }
        purpose: { type: string, description: 'What it proves: "Proof of death", "Proof of claim".' }
        from:
          type: object
          required: [label]
          properties:
            partyId: { type: [string, 'null'], format: uuid }
            label: { type: string }
            detail: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/RequirementState' }
        stateNote: { type: [string, 'null'], description: 'Why it is in this state when that is not obvious: "TIN mismatch" for `not_enough`, "Under review: photocopy" for `received`.' }
        requestedAt: { type: string, format: date-time }
        followUpDays: { type: integer, description: Days until the follow-up deadline fires; 0 means never chased. }
        followUpCount: { type: integer }
        lastReminderAt: { type: [string, 'null'], format: date-time }
        receivedAt: { type: [string, 'null'], format: date-time }
        acceptedAt: { type: [string, 'null'], format: date-time }
        satisfiedBy: { type: [string, 'null'] }
        waivedAt: { type: [string, 'null'], format: date-time }
        waiveReason: { type: [string, 'null'] }
        version: { type: integer, format: int64 }
    RequirementList:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Requirement' } }
        nextCursor: { type: [string, 'null'] }
    Deadline:
      type: object
      required: [id, claimId, kind, what, dueAt, originalDueAt, state, attempt, fired, version]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/DeadlineKind' }
        requirementId: { type: [string, 'null'], format: uuid }
        documentId: { type: [string, 'null'], format: uuid, description: 'The document a `document_review_by` row is about.' }
        partyId: { type: [string, 'null'], format: uuid, description: 'The payee a `bank_details_by` or reissue row is about.' }
        what: { type: string }
        sla: { type: [string, 'null'], description: 'The service level this row enforces: ack, forms, contact, status, decide, review, pay, docReview, reissue.' }
        dueAt: { type: string, format: date-time }
        originalDueAt: { type: string, format: date-time }
        extensionReason: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/DeadlineState' }
        attempt: { type: integer, description: Dispatch attempts so far. }
        dispatchedAt: { type: [string, 'null'], format: date-time }
        fired: { type: boolean, description: True when it came due and a deadline workflow ran for it. }
        lastOutcome: { type: [string, 'null'], enum: [started, already_started, start_failed, completed, skipped, closed_early, null] }
        lastError: { type: [string, 'null'] }
        workflowId: { type: [string, 'null'], examples: ['deadline-3f2b...'] }
        closedAt: { type: [string, 'null'], format: date-time }
        closedBy: { type: [string, 'null'], enum: [workflow, intake, requirement, decision, payment, user, null] }
        result: { type: [string, 'null'] }
        version: { type: integer, format: int64 }
    DeadlineList:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Deadline' } }
        nextCursor: { type: [string, 'null'] }
...
    Letter:
      type: object
      required: [id, claimId, templateCode, channel, recipientLabel, subject, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        benefitLineId: { type: [string, 'null'], format: uuid }
        templateCode: { type: string, examples: [ACK-LIFE-01] }
        channel: { type: string, enum: [letter, email, sms, portal, fax] }
        recipientPartyId: { type: [string, 'null'], format: uuid }
        recipientLabel: { type: string }
        subject: { type: string }
        summary: { type: [string, 'null'] }
        status:
          type: string
          enum: [draft, queued, sent, failed, skipped]
          description: |
            `skipped` is new: the letter was queued but the situation changed before it went (for example a re-run of the document workflow accepted the
            W-9 that run 1 had asked to be corrected), so it will never be sent; `statusReason` says why. A workflow that has finished leaves no letter `queued`.
        statusReason: { type: [string, 'null'], description: 'Why the letter is `skipped`; null otherwise. Additive.' }
        sentAt: { type: [string, 'null'], format: date-time }
        workflowId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    HistoryEvent:
      type: object
      required: [id, claimId, occurredAt, recordedAt, type, title, actorKind, actor]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time, description: When it happened on the claim. }
        recordedAt: { type: string, format: date-time, description: When we wrote it. }
        type: { type: string, enum: [decision, data, document, communication, task, access, assistant, payment] }
        title: { type: string }
        actorKind: { type: string, enum: [user, system, workflow, batch, portal], description: Structured actor; the mock packed this into a free-text `actor` string. }
        actor:
          type: string
          description: |
            Who did it, named the same way on every event: for `actorKind` `user`, the staff member's display name ("Rachel Kim") when the handle in `X-Actor` is an active user of `GET /staff`
            (any case), otherwise the handle as sent ("examiner.dev" when X-Actor was absent, "ops.dev"); for `system`, `workflow`, `batch` and `portal`, the label or workflow id. Events written before this rule may carry a raw handle.
        detail: { type: [string, 'null'] }
        ref: { type: [string, 'null'] }
        workflowId: { type: [string, 'null'] }
...
    RunStep:
      type: object
      required: [label, detail, system, state]
      properties:
        label: { type: string }
        detail: { type: string }
        system: { type: string, description: The system the step talks to. }
        state: { type: string, enum: [done, retried, skipped, failed], description: 'retried = Temporal ran the step again and it then succeeded; failed = the retries ran out (the run failed).' }
        attempts: { type: integer, minimum: 1, default: 1, description: 'How many times Temporal ran the step (1 = first time).' }
        note: { type: [string, 'null'], description: 'What Temporal did for this step, in plain words.' }
    WorkflowRun:
      type: object
      required: [id, workflowId, runId, type, name, startedBy, status, startedAt, steps, saved]
      properties:
        id: { type: string, format: uuid }
        workflowId: { type: string, examples: ['orch-L-26-043310-intake', 'deadline-3f2b...', 'event-...'] }
        runId: { type: string, description: Temporal's run id. }
        type: { type: string, enum: [orchestration, event, deadline] }
        name: { type: string }
        claimId: { type: [string, 'null'], format: uuid }
        startedBy: { type: string }
        triggerKind: { type: [string, 'null'], enum: [outbox_event, deadline, null] }
        triggerId: { type: [string, 'null'], format: uuid }
        status: { type: string, enum: [running, completed, skipped, needs_review, failed] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        steps: { type: array, items: { $ref: '#/components/schemas/RunStep' } }
        saved: { type: array, items: { type: string }, description: 'What the run wrote back to Postgres, in one transaction.' }
        error: { type: [string, 'null'] }
        runNo: { type: integer, minimum: 1, description: '1 for the first run of a workflow id; 2 for an ops re-run after it failed.' }
        rerunOfId: { type: [string, 'null'], format: uuid, description: 'For a re-run, the failed run it re-runs.' }
        note: { type: [string, 'null'], description: '"What Temporal did": the run''s story in plain words (retries, the failure, the re-run, the fault that caused it).' }
        elapsedMs: { type: [integer, 'null'], description: 'Real time the run took, from its first activity to its end.' }
        canRerun: { type: boolean, description: 'True for a failed run that is the latest of its workflow id: `POST /workflow-runs/{id}:rerun`.' }
...